Shortcut was missing batch normalisation
