+++ conf/models/segmentation/kpconv.yaml
@@ -168,7 +168,7 @@ models:
                     [True, True],
                 ]
             max_num_neighbors:
-                [[12, 12], [16, 16], [20, 20], [24, 24], [30, 30]]
+                [[12, 12], [25, 25], [38, 38], [36, 36], [32, 32]]
             module_name: KPDualBlock
         up_conv:
             module_name: FPModule
@@ -190,6 +190,6 @@ models:
                     bn_momentum,
                 ]
         mlp_cls:
-            nn: [in_feat, in_feat]
+            nn: [in_feat, 2*in_feat, 4* in_feat]
             dropout: 0.5
             bn_momentum: bn_momentum
