+++ src/core/base_conv/base_conv.py
@@ -1,22 +1,10 @@
 from abc import ABC
-import numpy as np
-import torch
+
+from src.core.common_modules.base_modules import BaseModule
 
 
-class BaseConvolution(ABC, torch.nn.Module):
+class BaseConvolution(ABC, BaseModule):
     def __init__(self, sampler, neighbour_finder, *args, **kwargs):
-        torch.nn.Module.__init__(self)
-
+        BaseModule.__init__(self)
         self.sampler = sampler
         self.neighbour_finder = neighbour_finder
-
-    @property
-    def nb_params(self):
-        """[This property is used to return the number of trainable parameters for a given layer]
-        It is useful for debugging and reproducibility.
-        Returns:
-            [type] -- [description]
-        """
-        model_parameters = filter(lambda p: p.requires_grad, self.parameters())
-        self._nb_params = sum([np.prod(p.size()) for p in model_parameters])
-        return self._nb_params
